+++ R/data.R
@@ -12,13 +12,13 @@
 #'     \item{\code{continent_residence}}{Continent of origin of population}
 #'     \item{\code{continent_origin}}{Continent of residence of population}
 #'   }
-#' @source UNHCR (The UN Refugee Agency) (\url{http://popstats.unhcr.org/en/overview})
+#' @source UNHCR (The UN Refugee Agency) (\url{https://www.unhcr.org/})
 "unhcr_popstats_2017"
 
 
 #' UNHCR data by continent of origin
 #'
-#' The dataset contains data about UNHCR's populations of concern summarised by continent of orginin.
+#' The dataset contains data about UNHCR's populations of concern summarised by continent of origin.
 #'
 #' @format A data frame with 913 observations on the following 4 variables.
 #'   \describe{
@@ -28,6 +28,6 @@
 #'     \item{\code{continent_origin}}{Continent of residence of population.}
 #'     \item{\code{n}}{Number of people concerned.}
 #'   }
-#' @source UNHCR (The UN Refugee Agency) (\url{http://popstats.unhcr.org/en/overview})
+#' @source UNHCR (The UN Refugee Agency) (\url{https://www.unhcr.org/})
 "unhcr_ts"