+++ NEWS.md
@@ -10,7 +10,7 @@ apexcharter 0.1.3
 apexcharter 0.1.2
 ==================
 
-* Update ApexCharts to 3.8.2
+* Upgraded ApexCharts.js to 3.8.2
 * Set parent container height to 0 by default (fix [#2](https://github.com/dreamRs/apexcharter/issues/2)).